backend/requirements.txt
# Core
Django==4.2.29
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Auth
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Async tasks
celery==5.3.6
redis==5.0.8
django-celery-beat==2.6.0

# Admin theme
django-jazzmin==2.6.0

# Filtering
django-filter==23.5

# Storage
django-storages==1.14.4
boto3==1.34.0

# Validation & utils
phonenumbers==8.13.44
python-decouple==3.8
Pillow==10.4.0

# NF-e / HTTP
requests==2.32.3

# Dev
ipython==8.12.3

# Prod
gunicorn==22.0.0
whitenoise==6.7.0
reportlab==4.2.5
